+++ angular/src/lib/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.1.3](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.1.2...@momentum-ui/angular@8.1.3) (2019-08-16)
+
+**Note:** Version bump only for package @momentum-ui/angular
+
+
+
+
+
+## [8.1.2](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.1.1...@momentum-ui/angular@8.1.2) (2019-08-14)
+
+
+### Bug Fixes
+
+* **toggle-switch:** correct registerOnChange call, add setDisabledState, emitter no longer going undefined. ([52b8594](https://github.com/momentum-design/momentum-ui/commit/52b8594))
+
+
+
+
+
 ## [8.1.1](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.0.4...@momentum-ui/angular@8.1.1) (2019-08-12)
 
 **Note:** Version bump only for package @momentum-ui/angular