+++ angularjs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [12.2.3](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angularjs@12.2.2...@momentum-ui/angularjs@12.2.3) (2019-08-16)
+
+**Note:** Version bump only for package @momentum-ui/angularjs
+
+
+
+
+
 ## [12.2.2](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angularjs@12.2.1...@momentum-ui/angularjs@12.2.2) (2019-08-14)
 
 **Note:** Version bump only for package @momentum-ui/angularjs